Adds category products view
Introduces a new page to display products belonging to a specific category.
This includes fetching the category name and associated products from
the API, and displaying them in a grid layout.
Clicking on a product will eventually lead to adding it to the order ticket.

Also, navigates to the CategoryProducts view when clicking on a category
in the NewOrder page.